// src/components/EndUserComm/http-services-sdk/be-types.ts
export enum Alignment { /* com.aternity.endusercomm.message.Alignment */
    Left = 'Left',
    Center = 'Center',
    Right = 'Right',
}
// ...
export enum MessageType { /* com.aternity.endusercomm.message.MessageType */
    Notification = 'Notification',
}

export interface MessageButton {// com.aternity.endusercomm.message.MessageButton, depends on classes: com.aternity.endusercomm.message.ButtonAction, com.aternity.endusercomm.message.ButtonStyle
    text: string
    style: ButtonStyle
    action: ButtonAction
}

export interface MessageButtons {// com.aternity.endusercomm.message.MessageButtons, depends on classes: com.aternity.endusercomm.message.Alignment, com.aternity.endusercomm.message.MessageButton
    alignment: Alignment
    buttons: MessageButton[]
}

export interface MessageButtonsText {// com.aternity.endusercomm.message.MessageButtonsText
    buttonText: string[]
}

export interface MessageChannel {// com.aternity.endusercomm.message.MessageChannel
    riverbed: boolean
}

export interface MessageLogo {// com.aternity.endusercomm.message.MessageLogo, depends on classes: com.aternity.endusercomm.message.Alignment
    alignment: Alignment
}

export interface UiBrandingInfo {// com.aternity.endusercomm.management.UiBrandingInfo
    brandingId: string
    brandingName: string
    primaryColor: string
    secondaryColor: string
    backgroundColor: string
    textColor: string
    logo: string
}

export interface UiMessageTemplate {// com.aternity.endusercomm.management.UiMessageTemplate, depends on classes: com.aternity.endusercomm.management.UiBrandingInfo, com.aternity.endusercomm.message.MessageButtons, com.aternity.endusercomm.message.MessageChannel, com.aternity.endusercomm.message.MessageLogo, com.aternity.endusercomm.message.MessageType
    id: string
    name: string
    description: string
    updatedAt: number
    updatedBy: string
    baseTemplateName: string
    type: MessageType
    channels: MessageChannel
    html: string
    css: string
    titleLexical: string
    bodyLexical: string
    branding: UiBrandingInfo
    buttons: MessageButtons
    logo: MessageLogo
}
// ...
export interface UiMessageTemplateInfo {// com.aternity.endusercomm.management.UiMessageTemplateInfo, depends on classes: com.aternity.endusercomm.message.MessageChannel, com.aternity.endusercomm.message.MessageType
    id: string
    name: string
    updatedAt: number
    description: string
    type: MessageType
    channels: MessageChannel
}